use test projects for spring-boot tests
