+++ jib-cli/src/integration-test/resources/gradle/projects/springBootLayered/build.gradle
@@ -1,0 +1,32 @@
+plugins {
+    id 'org.springframework.boot' version '2.3.7.RELEASE'
+    id 'io.spring.dependency-management' version '1.0.10.RELEASE'
+    id 'java'
+}
+
+//tasks.wrapper.configure {
+//    gradleVersion = "5.6.4"
+//}
+
+//group = 'com.example'
+//version = '0.0.1-SNAPSHOT'
+sourceCompatibility = '1.8'
+
+bootJar {
+    layered {
+        enabled = true
+    }
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+}
+
+test {
+    useJUnitPlatform()
+}
+
