+++ jib-cli/src/integration-test/resources/gradle/projects/springBootLayered/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'springboot-layered'
