--- jib-cli/src/integration-test/resources/gradle/projects/springbootLayered/build.gradle
@@ -1,22 +1,0 @@
-plugins {
-    id 'org.springframework.boot' version '2.1.6.RELEASE'
-    id 'io.spring.dependency-management' version '1.0.6.RELEASE'
-    id 'java'
-}
-
-//group = 'com.example'
-//version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '1.8'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-}
-
-test {
-    useJUnitPlatform()
-}
-
