--- jib-cli/src/integration-test/resources/gradle/projects/springbootLayered/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'springboot-layered'
