--- jib-cli/src/integration-test/resources/gradle/projects/springbootLayered/src/main/java/hello/HelloController.java
@@ -1,13 +1,0 @@
-package hello;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloController {
-
-  @RequestMapping("/")
-  public String index() {
-    return "Hello world";
-  }
-}
